admin role

+++ internal/entity/product.go
@@ -12,8 +12,8 @@ type Product struct {
 	Name        string                 `json:"name" gorm:"column:name"`
 	Description string                 `json:"descr" gorm:"column:description"`
 	Category    string                 `json:"category" gorm:"column:category"`
-	Price       uint64                 `json:"price" gorm:"column:price"`
-	Count       uint64                 `json:"count" gorm:"column:count"`
+	Price       uint                   `json:"price" gorm:"column:price"`
+	Count       uint                   `json:"count" gorm:"column:count"`
 	Active      bool                   `json:"active" gorm:"column:active"`
 	Options     map[string]interface{} `json:"options" gorm:"column:options;type:json"`
 	Article     string                 `json:"article" gorm:"column:article;unique"`
@@ -28,8 +28,8 @@ type CreateProduct struct {
 	Name        string                 `json:"name"`
 	Description string                 `json:"descr"`
 	Category    string                 `json:"category"`
-	Price       uint64                 `json:"price"`
-	Count       uint64                 `json:"count"`
+	Price       uint                   `json:"price"`
+	Count       uint                   `json:"count"`
 	Active      bool                   `json:"active"`
 	Options     map[string]interface{} `json:"options,omitempty"`
 }
@@ -38,8 +38,8 @@ type UpdateProduct struct {
 	Name        string                 `json:"name,omitempty"`
 	Description string                 `json:"descr,omitempty"`
 	Category    string                 `json:"category,omitempty"`
-	Price       uint64                 `json:"price,omitempty"`
-	Count       uint64                 `json:"count,omitempty"`
+	Price       uint                   `json:"price,omitempty"`
+	Count       uint                   `json:"count,omitempty"`
 	Active      bool                   `json:"active,omitempty"`
 	Options     map[string]interface{} `json:"options,omitempty"`
 }
